change test setting to increase stability
